simple-mcp-server/samples/langgraph-mcp-agent/pyproject.toml
[project]
name = "neo4j-mcp-agent"
version = "0.1.0"
description = "LangGraph ReAct agent for Neo4j MCP server using Azure OpenAI"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=1.0.5",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai>=0.3.0",
    "httpx>=0.27.0",
    "langchain>=1.2.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "azure-identity>=1.15.0",
]
